pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "cryptography>=46.0.3",
    "fastapi>=0.120.0",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "python-socketio>=5.14.2",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
]
